verificação dos exercicios feitos

+++ 1-Semestre/Algoritmos/Lista04-10Set/exercicio_46.py
@@ -1,4 +1,4 @@
-saldo = int(input('\nDigite o seu saldo: '))
+saldo = float(input('\nDigite o seu saldo: '))
 nsaldo = saldo * 1.01
 print(f'\nNovo salvo: {nsaldo}')
 print('\n')
