+++ 1-Semestre/Algoritmos/Lista04-10Set/exercicio_48.py
@@ -1,7 +1,7 @@
 import math as m
 
-sm = int(input('\nEntre com o salário mínimo: '))
-quantidade = int(input('\nEntre com a quantidade de quilowatt: '))
+sm = float(input('\nEntre com o salário mínimo: '))
+quantidade = float(input('\nEntre com a quantidade de quilowatt: '))
 preco = sm / 700
 vp = preco * quantidade
 vd = vp * 0.9
