+++ 1-Semestre/Algoritmos/Lista04-10Set/exercicio_51.py
@@ -1,4 +1,4 @@
-raio = int(input('\nDigite o raio: '))
+raio = float(input('\nDigite o raio: '))
 perimetro = 2* 3.14 * raio
 area = 3.14 * raio ** 2
 print(f'\nPerimetro: {perimetro} ')
