+++ 1-Semestre/Algoritmos/Lista04-10Set/exercicio_65.py
@@ -1,5 +1,5 @@
-altura = int(input('\nDigite a altura da lata: '))
-raio = int(input('\nDigite o raio da lata: '))
+altura = float(input('\nDigite a altura da lata: '))
+raio = float(input('\nDigite o raio da lata: '))
 volume = 3.14 * raio ** 2 * altura
 print(f'\nO volume da lata é: {volume}')
 print('\n')
