+++ 1-Semestre/Algoritmos/Lista04-10Set/exercicio_66.py
@@ -1,5 +1,5 @@
-tempo = int(input('\nDigite o tempo gasto: '))
-vel = int(input('\nDigite a velocidade média: '))
+tempo = float(input('\nDigite o tempo gasto: '))
+vel = float(input('\nDigite a velocidade média: '))
 dist = tempo * vel
 litros = dist / 12
 print(f'\nVelocidade: {vel} \nTempo: {tempo} \nDistância: {dist} \nLitros: {litros}')
