+++ 1-Semestre/Algoritmos/Lista04-10Set/exercicio_69.py
@@ -1,4 +1,4 @@
-num = float(input('\nDigite um numerador: '))
-denom = float(input('\nDigite denominador: '))
+num = int(input('\nDigite um numerador: '))
+denom = int(input('\nDigite denominador: '))
 print(f'\nDecimal: {num / denom}')
 print('\n')
