+++ 1-Semestre/Algoritmos/Lista04-10Set/exercicio_77.py
@@ -1,5 +1,5 @@
-a = int(input('\nDigite um número: '))
-b = int(input('\nDigite outro número: '))
+a = float(input('\nDigite um número: '))
+b = float(input('\nDigite outro número: '))
 d = (a - b) ** 2
 q = a**2 - b**2
 print(f'\nO quadrado da diferenca {d} \nDiferenca dos quadrados: {q}')
