+++ 1-Semestre/Algoritmos/Lista04-10Set/exercicio_78.py
@@ -1,4 +1,4 @@
 n = int(input('\nDigite o numero de lados do poligono: '))
-nd = n * ( n - 3) / 2
+nd = float(n * ( n - 3) / 2)
 print(f'\nNumero de diagonais: {nd}')
 print('\n')
